Don't hijack the caught signal completely
Previously after the backtrace log was written by the module it would simply
exit the process, which means that the signal is not processed further by the
system and nothing is written to syslog (which is bad for issue monitoring).

Changed this to not exit the process so the system handles the signal as usual,
and this module only adds the writing of backtrace log on top of that
(which imho is the expected behavior).